requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.23
langchain==0.0.335
langchain-community==0.0.13
ollama==0.1.4
python-dotenv==1.0.0
markdown==3.5.1
beautifulsoup4==4.12.2
requests==2.31.0 
torch==2.2.2
torchaudio==2.2.2
soundfile==0.12.1
chattts==0.2.3
huggingface_hub==0.22.0